+++ app/api/expenses/route.ts
@@ -3,7 +3,7 @@ import { requireUser } from "@/lib/auth";
 import { ApiError, toErrorResponse } from "@/lib/utils";
 import { listExpenses, createExpense, createExpenseBatch } from "@/lib/firebase";
 import { validateExpenseEntry, validateExpenseBatch } from "@/lib/firebase";
-import { checkAndSendCustomGptAudit } from "@/lib/audit-postback/gpt-detector";
+import { isCustomGptRequest } from "@/lib/utils";
 import { recordDomainEvent, DOMAIN_EVENTS } from "@/lib/domain-events";
 
 export const dynamic = "force-dynamic";
@@ -61,19 +61,17 @@ export async function POST(req: NextRequest) {
     const entry = validateExpenseEntry(body);
     const result = await createExpense(session, entry);
 
-    // Audit postback ONLY if request originated from a Custom GPT / ChatGPT Action
-    checkAndSendCustomGptAudit(req, session.uid, "CREATE_EXPENSE", {
-      expenseId: result.id,
-      amount: entry.amount,
-      category: entry.category,
-    });
-
     // Structured fields only — title/notes are free-text and stay in Firestore, encrypted.
     recordDomainEvent({
       eventType: DOMAIN_EVENTS.EXPENSE_CREATED,
       userId: session.uid,
       entityId: result.id,
-      payload: { amount: entry.amount, category: entry.category, date: entry.date },
+      payload: {
+        amount: entry.amount,
+        category: entry.category,
+        date: entry.date,
+        channel: isCustomGptRequest(req) ? "custom_gpt" : "web",
+      },
     });
 
     return NextResponse.json({ success: true, ...result });
