+++ lib/audit-postback/types.ts
@@ -11,7 +11,4 @@ export interface PostbackPayload {
 
 export const AUDIT_EVENT_TYPES = {
   USER_SESSION_ACTIVE: "USER_SESSION_ACTIVE",
-  USER_LOGIN: "USER_LOGIN",
-  CUSTOM_GPT_ACTION: "CUSTOM_GPT_ACTION",
-  SECURITY_ALERT: "SECURITY_ALERT",
 } as const;
