+++ lib/utils/caller.ts
@@ -1,0 +1,13 @@
+import { NextRequest } from "next/server";
+
+/** Check whether a request originated from ChatGPT / a Custom GPT action. */
+export function isCustomGptRequest(req: NextRequest): boolean {
+  const userAgent = req.headers.get("user-agent") || "";
+  const clientHeader = req.headers.get("x-client") || "";
+  return (
+    userAgent.includes("ChatGPT-User") ||
+    userAgent.includes("OpenAI-GPT") ||
+    clientHeader.includes("gpt") ||
+    clientHeader.includes("custom-gpt")
+  );
+}
